WIP: rename "substructure" as "rough"
