+++ src/plcc/rough_parser/__init__.py
@@ -2,4 +2,4 @@ from .parse_lines import Line, parse_lines
 from .parse_blocks import Block, parse_blocks, UnclosedBlockError
 from .parse_includes import Include, parse_includes
 from .parse_dividers import Divider, parse_dividers
-from .parse_substructure import parse_substructure
+from .parse_rough import parse_rough
