+++ src/plcc/rough_parser/process_includes_test.py
@@ -3,7 +3,7 @@ from pytest import raises, mark, fixture
 
 from .parse_lines import parse_lines, Line
 from .parse_includes import parse_includes
-from .load_substructure import process_includes, CircularIncludeError
+from .load_rough import process_includes, CircularIncludeError
 
 
 def test_None_yields_nothing():
